requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap3==2024.1
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
django-parler==2.3
django-rosetta==0.10.0
idna==3.6
pillow==10.2.0
polib==1.2.0
requests==2.31.0
requests-file==2.0.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
whitenoise==6.6.0
